fix: 자료형 수정

+++ src/main/java/com/example/enstart/model/SubscriptionId.java
@@ -15,8 +15,8 @@ import java.io.Serializable;
 public class SubscriptionId implements Serializable {
 
     @Column(name = "subscriber_id")
-    private Long subscriberId;
+    private Integer subscriberId;
 
     @Column(name = "service_id")
-    private Long serviceId;
+    private Integer serviceId;
 }
